acne-detection/requirements.txt
tensorflow-gpu==2.6.0
numpy==1.19.5
pandas==1.3.5
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
opencv-python==4.5.5.64
jupyterlab==3.2.9
